+++ src/entities/FarmerEntities.ts
@@ -53,6 +53,9 @@ export interface GetFarmerEntity {
   address: AddressEntity;
   farmerPlot: FarmerPlotEntity[];
   file: ImageEntity[];
+  createdAt?: string;
+  createBy?: string;
+  comment?: string;
 }
 export const GetFarmerEntity_INIT: GetFarmerEntity = {
   id: "",
@@ -67,6 +70,7 @@ export const GetFarmerEntity_INIT: GetFarmerEntity = {
   address: AddressEntity_INIT,
   farmerPlot: [FarmerPlotEntity_INIT],
   file: [ImageEntity_INTI],
+  comment: "",
 };
 
 export interface FarmerPageEntity {
@@ -87,6 +91,7 @@ export interface CreateFarmerEntity {
   birthDate: string;
   address: CreateAddressEntity;
   farmerPlot: FarmerPlotEntity[];
+  comment?: string;
 }
 export const CreateFarmerEntity_INIT: CreateFarmerEntity = {
   firstname: "",