+++ src/pages/droner/AddDroner/index.tsx
@@ -65,6 +65,8 @@ import { LAT_LNG_BANGKOK } from "../../../definitions/Location";
 import GoogleMap from "../../../components/map/GoogleMap";
 import moment from "moment";
 import locale from "antd/es/date-picker/locale/th_TH";
+import TextArea from "antd/lib/input/TextArea";
+import { useLocalStorage } from "../../../hook/useLocalStorage";
 const dateFormat = "DD/MM/YYYY";
 const dateCreateFormat = "YYYY-MM-DD";
 
@@ -72,7 +74,7 @@ const { Map } = require("immutable");
 const { Option } = Select;
 function AddDroner() {
   const [form] = Form.useForm();
-
+  const [profile] = useLocalStorage("profile", []);
   const [data] = useState<CreateDronerEntity>(
     CreateDronerEntity_INIT
   );
@@ -376,6 +378,7 @@ function AddDroner() {
       dronerArea,
       checkPlantsOther,
       idNo,
+      comment,
       ...rest
     } = form.getFieldsValue();
     const expPlant = [];
@@ -438,6 +441,7 @@ function AddDroner() {
         ...pushOtherPlant.toJS().dronerArea,
         mapUrl: values.mapUrl,
       },
+      createBy: `${profile?.firstname} ${profile?.lastname}`,
     };
     await DronerDatasource.createDronerList(payload).then(
       async (res) => {
@@ -1065,6 +1069,12 @@ function AddDroner() {
               />
             </Form.Item>
           </div>
+          <div className="form-group col-lg-12">
+            <label>หมายเหตุ</label>
+            <Form.Item name="comment">
+              <TextArea />
+            </Form.Item>
+          </div>
         </Form>
       </CardContainer>
     </div>
@@ -1183,6 +1193,7 @@ function AddDroner() {
       )}
       {showEditModal && (
         <ModalDrone
+          isEdit
           show={showEditModal}
           backButton={() => setShowEditModal((prev) => !prev)}
           callBack={insertDroneList}